refactoring: move common code of arm and avr implementation into a separate/new file
